server/requirements.txt
skyfield
python-decouple
pyowm